+++ modules/ustk_core/src/usImagePostScan3D.cpp
@@ -62,7 +62,7 @@ usImagePostScan3D::usImagePostScan3D() : usImage3D<unsigned char>(), usImageSett
 * @param isImageConvex Boolean to specyfy if the image was acquired by a convex probe(true) or by a linear probe (false).
 * @param isMotorConvex Boolean to specyfy if the image was acquired by a rotating  motor(true) or by a linear motor (false).
 */
-usImagePostScan3D::usImagePostScan3D(unsigned int AN, unsigned int LN, unsigned int FN, float probeRadius, float motorRadius, float scanLinePitch, float framePitch, bool isImageConvex, bool isMotorConvex)
+usImagePostScan3D::usImagePostScan3D(unsigned int AN, unsigned int LN, unsigned int FN, double probeRadius, double motorRadius, double scanLinePitch, double framePitch, bool isImageConvex, bool isMotorConvex)
   : usImage3D<unsigned char>(), usImageSettings3D(probeRadius, motorRadius, scanLinePitch, framePitch, isImageConvex, isMotorConvex)
 {
 
@@ -115,22 +115,22 @@ usImagePostScan3D::~usImagePostScan3D() {}
 * Setter for width Resolution.
 * @param widthResolution Width resolution (in meters) to set.
 */
-void usImagePostScan3D::setWidthResolution(float widthResolution) { m_widthResolution = widthResolution; }
+void usImagePostScan3D::setWidthResolution(double widthResolution) { m_widthResolution = widthResolution; }
 
 /**
 * Getter for width Resolution.
 * @return widthResolution Width resolution (in meters).
 */
-float usImagePostScan3D::getWidthResolution() { return m_heightResolution; }
+double usImagePostScan3D::getWidthResolution() { return m_heightResolution; }
 
 /**
 * Setter for width Resolution.
 * @param heightResolution Height resolution (in meters) to set.
 */
-void usImagePostScan3D::setHeightResolution(float heightResolution) { m_heightResolution = heightResolution; }
+void usImagePostScan3D::setHeightResolution(double heightResolution) { m_heightResolution = heightResolution; }
 
 /**
 * Setter for width Resolution.
 * @param heightResolution Height resolution (in meters) to set.
 */
-float usImagePostScan3D::getHeightResolution() { return m_heightResolution; }
+double usImagePostScan3D::getHeightResolution() { return m_heightResolution; }
